requirements.txt
absl-py==1.3.0
appdirs==1.4.4
astunparse==1.6.3
audioread==3.0.0
cachetools==5.2.0
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
contourpy==1.0.6
cycler==0.11.0
dataclasses==0.6
decorator==5.1.1
filelock==3.8.0
Flask==2.2.2
flatbuffers==22.10.26
fonttools==4.38.0
gast==0.5.3
google-api-core==2.10.2
google-auth==2.14.1
google-auth-oauthlib==0.4.6
google-cloud-aiplatform==1.19.0
google-cloud-bigquery==2.34.4
google-cloud-core==2.3.2
google-cloud-resource-manager==1.6.3
google-cloud-storage==2.6.0
google-crc32c==1.5.0
google-pasta==0.2.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpc-google-iam-v1==0.12.4
grpcio==1.50.0
grpcio-status==1.48.2
h5py==3.7.0
herpetologist==0.0.9
huggingface-hub==0.11.0
idna==3.4
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.2.0
keras==2.8.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.4
libclang==14.0.6
librosa==0.9.1
llvmlite==0.39.1
malaya-boilerplate==0.0.23
malaya-speech==1.2.7
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
memoization==0.4.0
noisereduce==2.0.1
numba==0.56.4
numpy==1.22.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==21.3
pandas==1.1.5
Pillow==9.3.0
pooch==1.6.0
proto-plus==1.22.1
protobuf==3.20.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyparsing==3.0.9
python-dateutil==2.8.2
python-speech-features==0.6
pytz==2022.6
PyYAML==6.0
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
resampy==0.4.2
rsa==4.9
scikit-learn==1.1.3
scipy==1.9.3
six==1.16.0
soundfile==0.11.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.8.0
tensorflow-io-gcs-filesystem==0.27.0
termcolor==2.1.0
tf-estimator-nightly==2.8.0.dev2021122109
threadpoolctl==3.1.0
tokenizers==0.13.2
torch==1.13.0
torchlibrosa==0.0.7
torchvision==0.14.0
tornado==6.2
tqdm==4.64.1
transformers==4.24.0
typing_extensions==4.4.0
Unidecode==1.3.6
urllib3==1.26.12
webrtcvad==2.0.10
Werkzeug==2.2.2
wrapt==1.14.1
zipp==3.10.0
